refactor(aqua): move baked registry into mise
