--- crates/aqua-registry/build.rs
@@ -1,284 +1,0 @@
-use std::collections::{HashMap, HashSet};
-use std::env;
-use std::fs;
-use std::fs::File;
-use std::io::Write;
-use std::path::{Path, PathBuf};
-
-use aqua_registry_types::{AquaPackage, RegistryYaml};
-use eyre::{Result, WrapErr, eyre};
-use flate2::Compression;
-use flate2::write::ZlibEncoder;
-use serde_yaml::Value;
-
-fn main() -> Result<()> {
-    let out_dir = env::var("OUT_DIR").wrap_err("OUT_DIR environment variable must be set")?;
-    generate_baked_registry(&out_dir)?;
-    generate_registry_metadata(&out_dir)?;
-    Ok(())
-}
-
-#[derive(Debug)]
-struct PackageRegistry {
-    id: String,
-    content: Vec<u8>,
-    aliases: Vec<String>,
-}
-
-fn generate_baked_registry(out_dir: &str) -> Result<()> {
-    let files_dest_path = Path::new(out_dir).join("aqua_standard_registry_files.rs");
-    let aliases_dest_path = Path::new(out_dir).join("aqua_standard_registry_aliases.rs");
-    let packages_dir = Path::new(out_dir).join("aqua_standard_registry_packages");
-
-    let registry_file = find_registry_file()?;
-
-    println!("cargo:rerun-if-changed={}", registry_file.display());
-
-    let content = fs::read_to_string(&registry_file).wrap_err_with(|| {
-        format!(
-            "Failed to read aqua registry file {}",
-            registry_file.display()
-        )
-    })?;
-
-    let registry_yaml = serde_yaml::from_str::<RegistryYaml>(&content).wrap_err_with(|| {
-        format!(
-            "Failed to parse aqua registry file {} as RegistryYaml",
-            registry_file.display()
-        )
-    })?;
-    let registry_value = serde_yaml::from_str::<Value>(&content).wrap_err_with(|| {
-        format!(
-            "Failed to parse aqua registry file {}",
-            registry_file.display()
-        )
-    })?;
-    let package_values = registry_value
-        .get("packages")
-        .and_then(|packages| packages.as_sequence())
-        .ok_or_else(|| {
-            eyre!(
-                "Aqua registry file {} does not contain a packages list",
-                registry_file.display()
-            )
-        })?;
-    let registries = package_registries(&registry_yaml.packages, package_values)?;
-    if registries.is_empty() {
-        return Err(eyre!(
-            "Aqua registry file {} contains no packages",
-            registry_file.display()
-        ));
-    }
-
-    fs::create_dir_all(&packages_dir).wrap_err("Failed to create baked registry package dir")?;
-    fs::write(
-        files_dest_path,
-        registry_files_code(&registries, &packages_dir)?,
-    )
-    .wrap_err("Failed to write baked registry files")?;
-    fs::write(aliases_dest_path, registry_aliases_code(&registries))
-        .wrap_err("Failed to write baked registry aliases")?;
-    Ok(())
-}
-
-fn generate_registry_metadata(out_dir: &str) -> Result<()> {
-    let metadata_dest_path = Path::new(out_dir).join("aqua_standard_registry_metadata.rs");
-    let metadata_file = find_registry_metadata_file()?;
-
-    println!("cargo:rerun-if-changed={}", metadata_file.display());
-
-    let content = fs::read_to_string(&metadata_file).wrap_err_with(|| {
-        format!(
-            "Failed to read aqua registry metadata file {}",
-            metadata_file.display()
-        )
-    })?;
-    let metadata = serde_yaml::from_str::<Value>(&content).wrap_err_with(|| {
-        format!(
-            "Failed to parse aqua registry metadata file {}",
-            metadata_file.display()
-        )
-    })?;
-    let repository = string_field(&metadata, "repository").ok_or_else(|| {
-        eyre!(
-            "Aqua registry metadata file {} does not contain a repository",
-            metadata_file.display()
-        )
-    })?;
-    let tag = string_field(&metadata, "tag").ok_or_else(|| {
-        eyre!(
-            "Aqua registry metadata file {} does not contain a tag",
-            metadata_file.display()
-        )
-    })?;
-
-    fs::write(
-        metadata_dest_path,
-        format!("AquaRegistryMetadata {{ repository: {repository:?}, tag: {tag:?} }}"),
-    )
-    .wrap_err("Failed to write baked registry metadata")?;
-    Ok(())
-}
-
-fn package_registries(
-    packages: &[AquaPackage],
-    package_values: &[Value],
-) -> Result<Vec<PackageRegistry>> {
-    if packages.len() != package_values.len() {
-        return Err(eyre!(
-            "parsed aqua package count mismatch: RegistryYaml has {}, raw YAML has {}",
-            packages.len(),
-            package_values.len()
-        ));
-    }
-
-    let mut registries = Vec::new();
-    for (package, package_value) in packages.iter().zip(package_values) {
-        let Some(id) = canonical_package_id(package) else {
-            continue;
-        };
-        let content = package_msgpack_z(package)?;
-        let aliases = package_aliases(package_value);
-        registries.push(PackageRegistry {
-            id,
-            content,
-            aliases,
-        });
-    }
-    Ok(registries)
-}
-
-fn registry_files_code(registries: &[PackageRegistry], packages_dir: &Path) -> Result<String> {
-    let mut used_stems = HashMap::new();
-    let mut entries = registries
-        .iter()
-        .map(|registry| {
-            let stem = package_file_stem(&registry.id);
-            if let Some(other_id) = used_stems.insert(stem.clone(), registry.id.as_str()) {
-                return Err(eyre!(
-                    "baked aqua registry package filename collision for {other_id:?} and {:?}: {stem}",
-                    registry.id
-                ));
-            }
-            let filename = format!("{stem}.msgpack.z");
-            let path = packages_dir.join(filename);
-            let mut file = File::create(&path)
-                .wrap_err_with(|| format!("Failed to write baked package {}", path.display()))?;
-            file.write_all(&registry.content)
-                .wrap_err_with(|| format!("Failed to write baked package {}", path.display()))?;
-            Ok((registry.id.clone(), path))
-        })
-        .collect::<Result<Vec<_>>>()?;
-    entries.sort_by(|a, b| a.0.cmp(&b.0));
-
-    Ok(registry_bytes_map_code(&entries))
-}
-
-fn registry_bytes_map_code(entries: &[(String, PathBuf)]) -> String {
-    let mut code = String::from("HashMap::from([\n");
-    for (key, path) in entries {
-        code.push_str(&format!(
-            "    ({key:?}, include_bytes!({:?}).as_slice()),\n",
-            path.display().to_string()
-        ));
-    }
-    code.push_str("])");
-    code
-}
-
-fn package_file_stem(id: &str) -> String {
-    let mut hash = 0xcbf29ce484222325u64;
-    for byte in id.as_bytes() {
-        hash ^= u64::from(*byte);
-        hash = hash.wrapping_mul(0x100000001b3);
-    }
-    format!("{hash:016x}")
-}
-
-fn registry_aliases_code(registries: &[PackageRegistry]) -> String {
-    let canonical_ids = registries
-        .iter()
-        .map(|registry| registry.id.as_str())
-        .collect::<HashSet<_>>();
-    let mut aliases = HashMap::new();
-
-    for registry in registries {
-        for alias in &registry.aliases {
-            if alias != &registry.id && !canonical_ids.contains(alias.as_str()) {
-                aliases.insert(alias.clone(), registry.id.clone());
-            }
-        }
-    }
-
-    let mut entries = aliases.into_iter().collect::<Vec<_>>();
-    entries.sort_by(|a, b| a.0.cmp(&b.0));
-
-    registry_string_map_code(&entries)
-}
-
-fn registry_string_map_code(entries: &[(String, String)]) -> String {
-    let mut code = String::from("HashMap::from([\n");
-    for (key, value) in entries {
-        code.push_str(&format!("    ({key:?}, {value:?}),\n"));
-    }
-    code.push_str("])");
-    code
-}
-
-fn package_msgpack_z(package: &AquaPackage) -> Result<Vec<u8>> {
-    let packed = rmp_serde::to_vec_named(package).wrap_err("Failed to serialize aqua package")?;
-    let mut zlib = ZlibEncoder::new(Vec::new(), Compression::best());
-    zlib.write_all(&packed)?;
-    Ok(zlib.finish()?)
-}
-
-fn canonical_package_id(package: &AquaPackage) -> Option<String> {
-    package
-        .name
-        .clone()
-        .or_else(|| {
-            if package.repo_owner.is_empty() || package.repo_name.is_empty() {
-                None
-            } else {
-                Some(format!("{}/{}", package.repo_owner, package.repo_name))
-            }
-        })
-        .or_else(|| package.path.clone())
-}
-
-fn package_aliases(package: &Value) -> Vec<String> {
-    package
-        .get("aliases")
-        .and_then(|aliases| aliases.as_sequence())
-        .map(|aliases| {
-            aliases
-                .iter()
-                .filter_map(|alias| string_field(alias, "name"))
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn string_field(value: &Value, key: &str) -> Option<String> {
-    value.get(key)?.as_str().map(str::to_string)
-}
-
-fn find_registry_file() -> Result<PathBuf> {
-    registry_file("registry.yaml", "Registry file")
-}
-
-fn find_registry_metadata_file() -> Result<PathBuf> {
-    registry_file("metadata.json", "Registry metadata file")
-}
-
-fn registry_file(file_name: &str, description: &str) -> Result<PathBuf> {
-    let manifest_dir = env::var("CARGO_MANIFEST_DIR")
-        .wrap_err("CARGO_MANIFEST_DIR environment variable must be set")?;
-    let embedded = Path::new(&manifest_dir)
-        .join("aqua-registry")
-        .join(file_name);
-    if embedded.exists() {
-        return Ok(embedded);
-    }
-    Err(eyre!("{description} not found at {}", embedded.display()))
-}
