+++ crates/aqua-registry/src/registry.rs
@@ -1,8 +1,6 @@
 use crate::types::{AquaPackage, RegistryYaml};
 use crate::{AquaRegistryConfig, AquaRegistryError, CacheStore, RegistryFetcher, Result};
-use flate2::read::ZlibDecoder;
 use std::collections::HashMap;
-use std::io::Read;
 use std::path::PathBuf;
 use std::sync::LazyLock;
 use tokio::sync::Mutex;
@@ -37,63 +35,6 @@ pub struct NoOpCacheStore;
 #[derive(Debug, Clone)]
 pub struct FileCacheStore {
     cache_dir: PathBuf,
-}
-
-/// Metadata for the baked aqua registry snapshot.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct AquaRegistryMetadata {
-    pub repository: &'static str,
-    pub tag: &'static str,
-}
-
-/// Baked canonical registry packages (compiled into binary).
-pub static AQUA_STANDARD_REGISTRY_FILES: LazyLock<HashMap<&'static str, &'static [u8]>> =
-    LazyLock::new(|| include!(concat!(env!("OUT_DIR"), "/aqua_standard_registry_files.rs")));
-
-/// Baked aqua registry snapshot metadata (compiled into binary).
-pub static AQUA_STANDARD_REGISTRY_METADATA: AquaRegistryMetadata = include!(concat!(
-    env!("OUT_DIR"),
-    "/aqua_standard_registry_metadata.rs"
-));
-
-/// Baked alias-to-canonical package ID map (compiled into binary).
-static AQUA_STANDARD_REGISTRY_ALIASES: LazyLock<HashMap<&'static str, &'static str>> =
-    LazyLock::new(|| {
-        include!(concat!(
-            env!("OUT_DIR"),
-            "/aqua_standard_registry_aliases.rs"
-        ))
-    });
-
-/// Returns all package IDs from the baked-in aqua registry.
-pub fn package_ids() -> Vec<&'static str> {
-    AQUA_STANDARD_REGISTRY_FILES.keys().copied().collect()
-}
-
-fn baked_registry_file(package_id: &str) -> Option<&'static [u8]> {
-    if let Some(content) = AQUA_STANDARD_REGISTRY_FILES.get(package_id) {
-        return Some(*content);
-    }
-
-    AQUA_STANDARD_REGISTRY_ALIASES
-        .get(package_id)
-        .and_then(|canonical| AQUA_STANDARD_REGISTRY_FILES.get(*canonical))
-        .copied()
-}
-
-fn decode_baked_package(package_id: &str, bytes: &[u8]) -> Result<AquaPackage> {
-    let mut zlib = ZlibDecoder::new(bytes);
-    let mut packed = Vec::new();
-    zlib.read_to_end(&mut packed).map_err(|err| {
-        AquaRegistryError::RegistryNotAvailable(format!(
-            "failed to decompress baked aqua-registry package {package_id}: {err}"
-        ))
-    })?;
-    rmp_serde::from_slice(&packed).map_err(|err| {
-        AquaRegistryError::RegistryNotAvailable(format!(
-            "failed to decode baked aqua package {package_id}: {err}"
-        ))
-    })
 }
 
 impl AquaRegistry {
@@ -188,36 +129,9 @@ impl RegistryFetcher for DefaultRegistryFetcher {
             return Ok(serde_yaml::from_str(&contents)?);
         }
 
-        // Fall back to baked registry if enabled
-        if self.config.use_baked_registry
-            && let Some(content) = baked_registry_file(package_id)
-        {
-            log::trace!("reading baked-in aqua-registry for {package_id}");
-            return Ok(RegistryYaml {
-                packages: vec![decode_baked_package(package_id, content)?],
-            });
-        }
-
         Err(AquaRegistryError::RegistryNotAvailable(format!(
             "no aqua-registry found for {package_id}"
         )))
-    }
-
-    async fn fetch_package(&self, package_id: &str) -> Result<AquaPackage> {
-        if self.config.use_baked_registry
-            && !self.config.cache_dir.join(".git").exists()
-            && let Some(content) = baked_registry_file(package_id)
-        {
-            log::trace!("reading baked-in aqua package for {package_id}");
-            return decode_baked_package(package_id, content);
-        }
-
-        let registry = self.fetch_registry(package_id).await?;
-        registry
-            .packages
-            .into_iter()
-            .next()
-            .ok_or_else(|| AquaRegistryError::PackageNotFound(package_id.to_string()))
     }
 }
 
@@ -293,51 +207,5 @@ mod tests {
         assert!(!cache.is_fresh("test"));
         assert!(cache.store("test", b"data").is_ok());
         assert!(cache.retrieve("test").unwrap().is_none());
-    }
-
-    #[test]
-    fn test_baked_registry_package_lookup() {
-        let registry = baked_registry_file("01mf02/jaq").unwrap();
-        let package = decode_baked_package("01mf02/jaq", registry).unwrap();
-
-        assert_eq!(package.repo_owner, "01mf02");
-        assert_eq!(package.repo_name, "jaq");
-    }
-
-    #[test]
-    fn test_baked_registry_path_only_package_lookup() {
-        let registry = baked_registry_file("golang.org/x/perf/cmd/benchstat").unwrap();
-        let package = decode_baked_package("golang.org/x/perf/cmd/benchstat", registry).unwrap();
-
-        assert_eq!(
-            package.path.as_deref(),
-            Some("golang.org/x/perf/cmd/benchstat")
-        );
-    }
-
-    #[test]
-    fn test_baked_registry_metadata() {
-        assert_eq!(
-            AQUA_STANDARD_REGISTRY_METADATA.repository,
-            "aquaproj/aqua-registry"
-        );
-        assert!(!AQUA_STANDARD_REGISTRY_METADATA.tag.is_empty());
-        assert!(AQUA_STANDARD_REGISTRY_METADATA.tag.starts_with('v'));
-    }
-
-    #[test]
-    fn test_baked_registry_alias_lookup() {
-        let alias = "elijah-potter/harper/harper-ls";
-
-        assert!(!AQUA_STANDARD_REGISTRY_FILES.contains_key(alias));
-        assert_eq!(
-            AQUA_STANDARD_REGISTRY_ALIASES.get(alias).copied(),
-            Some("Automattic/harper/harper-ls")
-        );
-
-        let registry = baked_registry_file(alias).unwrap();
-        let package = decode_baked_package(alias, registry).unwrap();
-
-        assert_eq!(package.name.as_deref(), Some("Automattic/harper/harper-ls"));
     }
 }
