+++ crates/aqua-registry/src/template.rs
