+++ crates/aqua-registry/src/types.rs
@@ -1,5 +1,3 @@
-mod template;
-
 use expr::{Context, Environment, Program, Value};
 use eyre::{Result, eyre};
 use indexmap::IndexSet;
@@ -435,7 +433,7 @@ impl AquaPackage {
         ctx.extend(self.vars_ctx()?);
         ctx.extend(overrides.clone());
 
-        template::render(s, &ctx)
+        crate::template::render(s, &ctx)
     }
 
     fn vars_ctx(&self) -> Result<HashMap<String, String>> {
