+++ src/aqua/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod aqua_registry_wrapper;
+pub(crate) mod standard_registry;
